src/test/java/edu/umn/cs/csci3081w/project/model/RouteDataTest.java
package edu.umn.cs.csci3081w.project.model;

import static org.junit.jupiter.api.Assertions.assertEquals;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

public class RouteDataTest {

  /**
   * Setup deterministic operations before each test runs.
   */
  @BeforeEach
  public void setUp() {
    PassengerFactory.DETERMINISTIC = true;
    PassengerFactory.DETERMINISTIC_NAMES_COUNT = 0;
    PassengerFactory.DETERMINISTIC_DESTINATION_COUNT = 0;
    RandomPassengerGenerator.DETERMINISTIC = true;
  }

  /**
   * Test route data constructor with no parameters.
   */
  @Test
  public void testConstructorWithNoParameters() {
    RouteData testRouteData = new RouteData();
    assertEquals("", testRouteData.getId());
    assertEquals(0, testRouteData.getStops().size());
  }

}
